Project Doumentation

+++ src/documents/documents.service.ts
@@ -15,7 +15,7 @@ export class DocumentsService {
   ) {}
 
   // Create a new document
-  async create(createDocumentDto: CreateDocumentDto): Promise<Document> {
+  async   create(createDocumentDto: CreateDocumentDto): Promise<Document> {
     const document = this.documentsRepository.create(createDocumentDto);
     return this.documentsRepository.save(document);
   }
@@ -37,15 +37,22 @@ export class DocumentsService {
   // Update a document
   async update(id: number, updateDocumentDto: UpdateDocumentDto): Promise<Document> {
     const document = await this.findOne(id);
+    if (!document) {
+      throw new NotFoundException(`Document with ID ${id} not found`);
+    }
     Object.assign(document, updateDocumentDto);
     return this.documentsRepository.save(document);
   }
 
   // Delete a document
-  async remove(id: number): Promise<void> {
+  async remove(id: number) {
     const document = await this.findOne(id);
+    if (!document) {
+      throw new NotFoundException(`Document with ID ${id} not found`);
+    }
     // Optionally, remove the file from the filesystem if required
     unlinkSync(join(__dirname, '..', '..', document.path)); // Example: Delete the file from disk
     await this.documentsRepository.remove(document);
+    return {message:`Document with ID ${id} Delete`};
   }
 }
